Rename the IntakeLift to the Lift Subsystem
