+++ src/main/java/frc/robot/subsystems/lift/LiftConstants.java
@@ -6,13 +6,15 @@
 /*- of this project.                                                      */
 /*------------------------------------------------------------------------*/
 
+package frc.robot.subsystems.lift;
+
 /**
- * This package contains the implementation of the <code>IntakeLift</code> subsystem.
+ * This class contains the constants for the <i>Intake</i> <code>Lift</code> subsystem.
  *
  * <p>More detail ...
  *
  * @since 2026.0.0
- * @author first.stu
+ * @author first.brian Buzzell
  * @version 2026.0.0
  */
-package frc.robot.subsystems.intakelift;
+public class LiftConstants {}
